lint code and resolve comments
