+++ core/conversion/converters/converter_util.cpp
@@ -384,7 +384,7 @@ nvinfer1::ITensor* scalar_to_tensor(ConversionCtx* ctx, at::Scalar s) {
   return out;
 }
 
-nvinfer1::ITensor* add_expand_layer(ConversionCtx* ctx,  nvinfer1::ITensor* in, nvinfer1::Dims expandedDims) {
+nvinfer1::ITensor* add_expand_layer(ConversionCtx* ctx, nvinfer1::ITensor* in, nvinfer1::Dims expandedDims) {
   auto input_dims = in->getDimensions();
   TORCHTRT_CHECK(
       input_dims.nbDims <= expandedDims.nbDims,
